Skip child suites when using this.skip()
When using `this.skip()` in a before hook, it skips all the tests in
the current suite, but runs the tests in all the child suites.

This patch expands the fixtures in skip-[a]sync-before fixtures to
cover the more expected behavior, and corrects the handling of Pending
exceptions in runner.js accordingly.

Fixes #2819